skills-pr2/perceive_objects: use action result data instead of topic
The topic has the drawback that it is received and parsed all the time.
With point clouds being part of the message this takes long and causes
warnings for the thread taking too long.

To fix this, do not use the update_objects skill, but rather directly
call the action in an ActionJumpState to gain access to the result at
the end. Then parse the data from that message, which is done only once.

This removes the ability to bypass the perception action call.