hybris-2014: add workaround for secondary inspection fail
If picking up an object failed, no other object was picked up for
inspection from then on.

The problem is that pickup senses holding, which becomes "none" on
failure. This means that poss(inspect_object(N)) does not hold. This
leaves the interrupt in a stale state and prevents it from being
executed again. We fix this by checking the pre-condition in the
interrupt before actually executing inspection (and avoiding calling
trans).

Debugged with Christoph Schwering.